% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisConfiguration.R
\name{add_databaseConnection}
\alias{add_databaseConnection}
\title{Function to add new database connection in config file}
\usage{
add_databaseConnection(
  database,
  dbms,
  port = NULL,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema,
  resultsDatabaseSchema,
  configFile = "config.yml"
)
}
\arguments{
\item{database}{the database names hosting the OMOP data (ex. synpuf_110k). Can take multiple inputs}

\item{dbms}{the dbms used for the OMOP database (ex. postgresql, redshift, sql server). This
argument takes all dbms types that are compatible with DatabaseConnector. Can take multiple inputs based on the number of databases entered}

\item{port}{the port used for the OMOP databse. Can take multiple inputs based on the number of databases entered}

\item{cdmDatabaseSchema}{the schema that holds the cdm tables. Make sure your user has read access.
Can take multiple inputs based on the number of databases entered.}

\item{resultsDatabaseSchema}{the schema that hosts the result tables for this study. This schema should be separate from the
the results schema used by webAPI. Make sure your user has read and write access. Can take multiple inputs based on the number
of databases entered.}

\item{configFile}{the location of the config file used for the analysis}

\item{vocabularyDatabaseSchmea}{the schema that holds the vocabulary tables. Many times this is the same as the cdm schema.
Make sure your user has read access. Can take multiple inputs based on the number of databases entered.}
}
\description{
This function will add a new database connection into the config.yml file
which specifies the configurations of the ohdsi study. The function will
prompt a keyring input for the user, host and password.
}
