% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{cohort_diagnostics_cohort_relationship}
\alias{cohort_diagnostics_cohort_relationship}
\title{cohort relationship analysis for cohort diagnostics}
\usage{
cohort_diagnostics_cohort_relationship(
  connectionDetails,
  cdmDatabaseSchema = config::get("cdmDatabaseSchema"),
  targetGeneratedCohort,
  comparatorGeneratedCohorts,
  temporalCovariateSettings
)
}
\arguments{
\item{connectionDetails}{ConnectionDetails}

\item{cdmDatabaseSchema}{(character) Schema where the CDM data lives in the database}

\item{temporalCovariateSettings}{temporal covariate settings used for analysis}

\item{generatedCohort}{dependency object of generated cohort class that tracks cohort used in incidence
analysis}
}
\description{
cohort relationship analysis for cohort diagnostics
}
