% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{cohort_diagnostics_incidence}
\alias{cohort_diagnostics_incidence}
\title{Incidence analysis for cohort diagnostics}
\usage{
cohort_diagnostics_incidence(
  connectionDetails,
  cdmDatabaseSchema = config::get("cdmDatabaseSchema"),
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  vocabularyDatabaseSchema = config::get("vocabularyDatabaseSchema"),
  generatedCohort,
  cdmVersion = 5,
  firstOccurrenceOnly,
  washoutPeriod
)
}
\arguments{
\item{connectionDetails}{ConnectionDetails}

\item{cdmDatabaseSchema}{(character) Schema where the CDM data lives in the database}

\item{tempEmulationSchema}{the temporary table schema used in certain databases}

\item{vocabularyDatabaseSchema}{(character) Schema where the vocabulary tables lives in the database}

\item{generatedCohort}{dependency object of generated cohort class that tracks cohort used in incidence
analysis}

\item{cdmVersion}{the version of the cdm >= 5}

\item{firstOccurrenceOnly}{a logic toggle for first occurrence}

\item{washoutPeriod}{number of days for washout}
}
\value{
A dataframe with incidence calculations with several combinations
}
\description{
Incidence stratified by age, gender and index year used in cohort diagnostics
}
