% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{cohort_diagnostics_inclusion_rules}
\alias{cohort_diagnostics_inclusion_rules}
\title{Inclusion Rule analysis for cohort diangostics}
\usage{
cohort_diagnostics_inclusion_rules(
  connectionDetails,
  generatedCohort,
  databaseId = config::get("databaseName")
)
}
\arguments{
\item{connectionDetails}{ConnectionDetails}

\item{generatedCohort}{dependency object of generated cohort class that tracks cohort used in incidence
analysis}

\item{databaseId}{(character) identify which database is being used}
}
\value{
A list of dataframes retrieving inclusion rule states generated for the cohorts
}
\description{
Extract information on inclusion rules
}
