% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{cohort_diagnostics_temporal_characterization}
\alias{cohort_diagnostics_temporal_characterization}
\title{Temporal cohort characterization analysis for cohort diagnostics}
\usage{
cohort_diagnostics_temporal_characterization(
  connectionDetails,
  cdmDatabaseSchema = config::get("cdmDatabaseSchema"),
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  generatedCohort,
  temporalCovariateSettings
)
}
\arguments{
\item{connectionDetails}{ConnectionDetails}

\item{cdmDatabaseSchema}{(character) Schema where the CDM data lives in the database}

\item{tempEmulationSchema}{the temporary table schema used in certain databases}

\item{generatedCohort}{dependency object of generated cohort class that tracks cohort used in incidence
analysis}

\item{temporalCovariateSettings}{temporal covariate settings used for analysis}
}
\description{
Temporal cohort characterization analysis for cohort diagnostics
}
