% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{cohort_diagnostics_time_series}
\alias{cohort_diagnostics_time_series}
\title{Time Series analysis for cohort diagnostics}
\usage{
cohort_diagnostics_time_series(
  connectionDetails,
  cdmDatabaseSchema = config::get("cdmDatabaseSchema"),
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  generatedCohort,
  observationPeriodRange = NULL
)
}
\arguments{
\item{connectionDetails}{ConnectionDetails}

\item{cdmDatabaseSchema}{(character) Schema where the CDM data lives in the database}

\item{tempEmulationSchema}{the temporary table schema used in certain databases}

\item{generatedCohort}{dependency object of generated cohort class that tracks cohort used in incidence
analysis}

\item{observationPeriodRange}{an object stating the start and end of database observation}
}
\value{
A dataframe with incidence calculations with several combinations
}
\description{
Time series stratified by age group and gender in cohort diagnostics
}
