% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_setup.R
\name{create_execution_settings}
\alias{create_execution_settings}
\title{Create execution settings}
\usage{
create_execution_settings(
  active_database,
  connectionDetails = config::get("connectionDetails"),
  cdmDatabaseSchema = config::get("cdmDatabaseSchema"),
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
  vocabularyDatabaseSchema = config::get("vocabularyDatabaseSchema"),
  resultsDatabaseSchema = config::get("resultsDatabaseSchema"),
  cohortTableName = config::get("cohortTableName"),
  databaseId = config::get("databaseName"),
  studyName = config::get("studyName")
)
}
\arguments{
\item{active_database}{the database we want to implement the pipeline}

\item{connectionDetails}{ConnectionDetails}

\item{cdmDatabaseSchema}{(character) Schema where the CDM data lives in the database}

\item{tempEmulationSchema}{the temporary table schema used in certain databases}

\item{vocabularyDatabaseSchema}{(character) Schema where the vocabulary tables lives in the database}

\item{resultsDatabaseSchema}{(character) Schema where the cohorts and results live}

\item{cohortTableName}{(character) naming convention for cohort tables}

\item{databaseId}{(character) identify which database is being used}

\item{studyName}{(character) provide a name of the study}
}
\value{
returns executionSettings object
}
\description{
Creates execution settings object that defines inputs for database connection
}
