% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_project.R
\name{create_ohdsitargets_project}
\alias{create_ohdsitargets_project}
\title{Create example ohdsitargets project}
\usage{
create_ohdsitargets_project(path)
}
\arguments{
\item{path}{Folder where the new project should be created}
}
\description{
Create example ohdsitargets project
}
