% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{define_temporal_covariates}
\alias{define_temporal_covariates}
\title{Define temporal covariates. Gives default settings}
\usage{
define_temporal_covariates()
}
\description{
Define temporal covariates. Gives default settings
}
