% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_diagnostics.R
\name{diagnostics_database_meta}
\alias{diagnostics_database_meta}
\title{Get meta information about database}
\usage{
diagnostics_database_meta(
  connectionDetails,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema,
  databaseId
)
}
\arguments{
\item{connectionDetails}{ConnectionDetails}

\item{cdmDatabaseSchema}{(character) Schema where the CDM data lives in the database}

\item{vocabularyDatabaseSchema}{(character) Schema where the vocabulary tables lives in the database}

\item{databaseId}{(character) identify which database is being used}
}
\value{
A tibble with meta information about the database
}
\description{
Retrieve database meta data
}
