% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_cohort}
\alias{generate_cohort}
\title{Generate one cohort}
\usage{
generate_cohort(
  connectionDetails,
  cdmDatabaseSchema,
  cohortTableRef,
  cohortId,
  cohortName,
  cohortFile,
  tempEmulationSchema = getOption("sqlRenderTempEmulationSchema")
)
}
\arguments{
\item{connectionDetails}{ConnectionDetails}

\item{cdmDatabaseSchema}{(character) Schema where the CDM data lives in the database}

\item{cohortTableRef}{A cohort table reference as returned by the \code{create_cohort_tables()} function}

\item{cohortId}{(integer) The id that should be assigned to the generated cohort}

\item{cohortName}{(character) The name of cohort}

\item{cohortFile}{(character) The path to the cohort json file relative to the project or working directory.}
}
\value{
An R object representing the generated cohort.
}
\description{
Generate a single cohort in the CDM database.
}
