% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generateCohort}
\alias{generateCohort}
\title{Generate one cohort}
\usage{
generateCohort(
  connectionDetails,
  cdmDatabaseSchema,
  cohortTableRef,
  cohortId,
  cohortName,
  cohortFile
)
}
\arguments{
\item{cohortTableRef}{A cohort table reference}

\item{cohortFile}{}
}
\description{
Generate one cohort
}
