% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisConfiguration.R
\name{initializeStudySession}
\alias{initializeStudySession}
\title{Function that initializes the study project for analysis}
\usage{
initializeStudySession(
  database,
  studyName,
  jarFolder = "jarFolder",
  diagnosticsFolder = "diagnostics",
  jsonFolder = "json",
  cohortOutputFolder = "cohort",
  dbms,
  server,
  user,
  port,
  cdmDatabaseSchema,
  vocabularyDatabaseSchema,
  resultsDatabaseSchema
)
}
\arguments{
\item{database}{the database names hosting the OMOP data (ex. synpuf_110k). Can take multiple inputs}

\item{studyName}{a name for this study; default is TREADS}

\item{jarFolder}{a folder to download the jdbc drivers for DatabaseConnector; default is jarFolder}

\item{diagnosticsFolder}{a folder to write CohortDiagnostics results; default is diagnostics}

\item{jsonFolder}{a folder to write cohort definition jsons; default is json}

\item{cohortOutputFolder}{a folder to write rds files for cohort definitions; default is cohort}

\item{dbms}{the dbms used for the OMOP database (ex. postgresql, redshift, sql server). This
argument takes all dbms types that are compatible with DatabaseConnector. Can take multiple inputs based on the number of databases entered}

\item{server}{the server name for the OMOP database. Contact your database admin for a full server string.
Can take multiple inputs based on the number of databases entered}

\item{user}{the user for the OMOP database. Can take multiple inputs based on the number of databases entered}

\item{port}{the port used for the OMOP databse. Can take multiple inputs based on the number of databases entered}

\item{cdmDatabaseSchema}{the schema that holds the cdm tables. Make sure your user has read access.
Can take multiple inputs based on the number of databases entered.}

\item{resultsDatabaseSchema}{the schema that hosts the result tables for this study. This schema should be separate from the
the results schema used by webAPI. Make sure your user has read and write access. Can take multiple inputs based on the number
of databases entered.}

\item{vocabularyDatabaseSchmea}{the schema that holds the vocabulary tables. Many times this is the same as the cdm schema.
Make sure your user has read access. Can take multiple inputs based on the number of databases entered.}
}
\description{
This function sets the config.yml file used to facilitate aspects of the
database connection used for the analysis. Contact your database administrator
to ensure you have all the information needed to run an OHDSI analysis. You must
have read access to the cdm and vocabulary schemas. You must have read and write
access to a user specific results schema (or a scratch schema). The results schema
should be seperate from the schema used by webAPI. This function does not have an input for
passwords, which is done by design. When this function is run, the user will be prompted to input
the password corresponding to the database user, using the {keyring} package in R. This input
does not leave a trace in the .Rhistory.
}
