% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisConfiguration.R
\name{pointToDatabase}
\alias{pointToDatabase}
\title{Function that orients the active database in the analysis}
\usage{
pointToDatabase(database)
}
\arguments{
\item{database}{the database for the analysis (ex. synpuf_110k)}
}
\description{
A study site can have multiple OMOP databases. Many times they use the same dbms, however,
the databases a separate. One must connect to each database in order to execute an analysis.
This functions helps the user to point to a specific database to conduct the analysis.
The database specified will become the active configuration for the config.yml file.
With the configuration specified, the correct environment variables corresponding
to the database will be retrieved to use for the analysis of that particular database.
This function will also create the study specific tables in the users results schema
to save cohort counts and use for cohort diagnostics. There is no return in this function
however, notes will be printed to the console informing the user of what environment
variables are being used.
}
