% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{tar_cohort_diagnostics}
\alias{tar_cohort_diagnostics}
\title{This is the target factory for cohort diagnostics}
\usage{
tar_cohort_diagnostics(
  cohortsToCreate,
  executionSettings,
  incidenceAnalysisSettings,
  temporalCovariateSettings
)
}
\arguments{
\item{cohortsToCreate}{a dataframe with the cohorts to create}

\item{executionSettings}{An object containing all information of the database connection created from config file}

\item{incidenceAnalysisSettings}{a dataframe where each row is a different combination of anlayis}

\item{temporalCovariateSettings}{a list of settings for cohort relationship analysis}
}
\description{
This is the target factory for cohort diagnostics
}
