% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{tar_cohort_incidence}
\alias{tar_cohort_incidence}
\title{Create references to inclusion rules analysis per cohort}
\usage{
tar_cohort_incidence(
  cohortsToCreate,
  incidenceAnalysisSettings,
  connectionDetails,
  active_database = "eunomia",
  cdmDatabaseSchema = config::get("cdmDatabaseSchema"),
  vocabularyDatabaseSchema = config::get("vocabularyDatabaseSchema"),
  databaseId = config::get("databaseName")
)
}
\arguments{
\item{cohortsToCreate}{A dataframe with one row per cohort and the following columns: cohortId, cohortName, cohortJsonPath}

\item{incidenceAnalysisSettings}{a dataframe where each row is a different combination of anlayis}

\item{connectionDetails}{ConnectionDetails}

\item{active_database}{the database used in the analysis}

\item{cdmDatabaseSchema}{(character) Schema where the CDM data lives in the database}

\item{vocabularyDatabaseSchema}{(character) Schema where the vocabulary tables lives in the database}

\item{databaseId}{(character) identify which database is being used}
}
\value{
One target for each generated cohort with names cohortInclusionRules_{id}
}
\description{
Creates targets of inclusion rules for each cohort in the project
}
