% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{tar_cohort_inclusion_rules}
\alias{tar_cohort_inclusion_rules}
\title{Create references to inclusion rules analysis per cohort}
\usage{
tar_cohort_inclusion_rules(
  cohortsToCreate,
  connectionDetails,
  active_database = "eunomia",
  databaseId = config::get("databaseName")
)
}
\arguments{
\item{cohortsToCreate}{A dataframe with one row per cohort and the following columns: cohortId, cohortName, cohortJsonPath}

\item{connectionDetails}{ConnectionDetails}

\item{active_database}{the database used in the analysis}

\item{databaseId}{(character) identify which database is being used}
}
\value{
One target for each generated cohort with names cohortInclusionRules_{id}
}
\description{
Creates targets of inclusion rules for each cohort in the project
}
