% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{tar_cohort_relationship}
\alias{tar_cohort_relationship}
\title{This is the target factory for the cohort relationships in cohort diagnostics}
\usage{
tar_cohort_relationship(
  cohortsToCreate,
  temporalCovariateSettings,
  executionSettings
)
}
\arguments{
\item{cohortsToCreate}{a dataframe with the cohorts to create}

\item{temporalCovariateSettings}{a list of settings for cohort relationship analysis}

\item{executionSettings}{An object containing all information of the database connection created from config file}
}
\description{
This is the target factory for the cohort relationships in cohort diagnostics
}
