% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{tar_cohort_tables}
\alias{tar_cohort_tables}
\title{Create cohort tables}
\usage{
tar_cohort_tables(
  name,
  connectionDetails = config::get("connectionDetails"),
  cohortDatabaseSchema = config::get("resultsDatabaseSchema"),
  cohortTableNames = config::get("cohortTableNames")
)
}
\arguments{
\item{name}{Symbol, the name of the cohort table target}

\item{connectionDetails}{Database connectionDetails object created by createConnectionDetails()}

\item{cohortDatabaseSchema}{character string with the schema where the cohort tables can be created}

\item{cohortTableNames}{A list of cohort table names created by cohortTableNames}
}
\value{
a target - list that contains the data needed to query the cohort table and the time it was created
}
\description{
Returns a target that references the cohort table in the database
}
