% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{tar_cohort_time_series}
\alias{tar_cohort_time_series}
\title{Creates targets of time series for each cohort in the project}
\usage{
tar_cohort_time_series(
  cohortsToCreate,
  executionSettings,
  observationPeriodRange = NULL
)
}
\arguments{
\item{cohortsToCreate}{A dataframe with one row per cohort and the following columns: cohortId, cohortName, cohortJsonPath}

\item{executionSettings}{An object containing all information of the database connection created from config file}

\item{observationPeriodRange}{the result of diagnostics_database_observation_period}
}
\value{
One target for each generated cohort with names cohortTimeSeries_{id}
}
\description{
Creates targets of time series for each cohort in the project
}
