% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{tar_cohorts}
\alias{tar_cohorts}
\title{Create references to generated cohorts}
\usage{
tar_cohorts(
  cohortsToCreate,
  connectionDetails = config::get("connectionDetails"),
  cdmDatabaseSchema = config::get("cdmDatabaseSchema"),
  cohortDatabaseSchema = config::get("resultsDatabaseSchema"),
  cohortTableName = config::get("cohortTableName")
)
}
\arguments{
\item{cohortsToCreate}{A dataframe with one row per cohort and the following columns: cohortId, cohortName, cohortJsonPath}

\item{name}{Symbol, base name for the collection of cohort targets}

\item{cohortTable}{A reference to the cohort table created by \code{tar_cohort_tables}}
}
\value{
One target for each cohort file and for each generated cohort with names cohort_{id}
}
\description{
Creates targets for each cohort in the project
}
