% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_diagnostics.R
\name{tar_cohort_incidence}
\alias{tar_cohort_incidence}
\title{Create references to inclusion rules analysis per cohort}
\usage{
tar_cohort_incidence(
  cohortsToCreate,
  incidenceAnalysisSettings,
  executionSettings
)
}
\arguments{
\item{cohortsToCreate}{A dataframe with one row per cohort and the following columns: cohortId, cohortName, cohortJsonPath}

\item{incidenceAnalysisSettings}{a dataframe where each row is a different combination of anlayis}

\item{executionSettings}{An object containing all information of the database connection created from config file}
}
\value{
One target for each generated cohort with names cohortInclusionRules_{id}
}
\description{
Creates targets of inclusion rules for each cohort in the project
}
