% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_diagnostics.R
\name{tar_database_diagnostics}
\alias{tar_database_diagnostics}
\title{Create series of targets that provide information about the database}
\usage{
tar_database_diagnostics(executionSettings)
}
\arguments{
\item{executionSettings}{An object containing all information of the database connection created from config file}
}
\value{
multiple targets with information about database (summarize more later)
}
\description{
Summarize information about the database used in the analysis
}
